Finished Task 4.1.1

+++ Tasks_4/4.1.1_File_Management_System/Data/Resource/Changer/ChangeResource.cs
@@ -11,14 +11,14 @@ namespace FileManagementSystem.Data.Resource.Changer
         {
             if (!Directory.Exists(pathFolder))
             {
-                throw new ArgumentException("Incorrect path!");
+                Directory.CreateDirectory(pathFolder);
             }
             PathResource = pathFolder;
         }
 
         public void ChangeAllResource(ResourceEventArgs[] args)
         {
-            if (Directory.Exists(PathResource))
+            if (Directory.Exists(PathResource) && args.Length > 0)
             {
                 Directory.Delete(PathResource, true);
             }
